+++ client/components/index.js
@@ -7,4 +7,5 @@ export {default as Navbar} from './navbar'
 export {default as UserHome} from './user-home'
 export {Login, Signup} from './auth-form'
 export {default as Main} from './main'
+
 export {default as Game} from './game'
