+++ public/style.css
@@ -112,3 +112,7 @@ li {
   -webkit-transform: scaleX(-1);
   transform: scaleX(-1);
 }
+.container-fluid {
+  background-color: green;
+  color: white;
+}
